updated the identifier ref for studies [fixes #2]
The identifier to reference simulation studies is the base64 encoding of
the URL to the model document. However, you'ver been using the default
base64 encoding, which may contain slashes! Lucky us, guess what happens
when you try to encode links to the biomodels database:

    echo "http://www.ebi.ac.uk/biomodels-main/download?mid=BIOMD00blahblah" | base64 -w 0

Thus, the generated URL to the detail view will look like:

    http://m2cat.sems.uni-rostock.de/model?id=something/somemore

which is not the problem yet. But that page will try to load the
detail-page loading status from

    http://m2cat.sems.uni-rostock.de/rest/result/something/somemore/status

The REST API, however, only recognises `.../result/{id}/status` and
doesn't answer the above request...

This commit changes the encoding and decoding behaviour to make sure the
encoded string is URL-safe!